config Heroku

+++ controllers/users.js
@@ -2,12 +2,19 @@ import UserModel from '../models/Users.js';
 
 export const createUser = async (req, res) => {
     try {
-        const newUser = new UserModel(req.body);
+        const data = req.body
+        // console.log(data.email)
+        // const checkEmail = checkIfUserExists(data.email)
 
-        await newUser.save();
-        res.status(200).send(`user created successfully`)
+        // if (!checkEmail) {
+            const newUser = new UserModel(data);
+            await newUser.save();
+            res.status(200).send(`user created successfully`)
+        // } else {
+        //     res.status(409).send(`E-mail cadastrado, tente novamente.`)
+        // }
     } catch (err) {
-        res.status(409).send({ message: err.message });
+        res.status(400).send({ message: err.message });
     }
 }
 
@@ -46,9 +53,22 @@ export const deleteUser = async (req, res) => {
     try {
         const user = await UserModel.findById(req.params.id);
 
-        user.deleteOne(user)
-        res.status(200).send("user deleted successfully")
+        if (user) {
+            user.deleteOne(user)
+            res.status(200).send("user deleted successfully")
+        } else {
+            res.status(404).send("Usuário não encontrado");
+        }
+
     } catch (err) {
         res.status(404).send({ message: err.message });
     }
 }
+
+// const checkIfUserExists = async (email) => {
+//     const emailExist = await UserModel.findOne({email})
+
+//     console.log(emailExist)
+//     return emailExist ? true : false;
+
+// }
